Add a script that insert open and close list tags
When copying and pasting ingredients and steps on a recipe website, I
end up with a bunch of lines that I need to add HTML list tags to.

Instead of doing this by hand, I wrote this  to automate the process. It
reads from an input file called "input.txt", skip the empty lines, add
the list tags to the rest and print the result.